Refactor Terms structure and load conditional C19 questions via flag
